refactor SourceDependencyAnalysis - Swap the class with its nested Analyzer class. Top-level is now   SourceDependencyAnalyzer; Results is now nested. - Move some nested types (JavaClass, JavaPackage, CompilationUnit,   and PackageDirectory) to top-level types.
